+++ modules/mcc/src/debug.cpp
@@ -14,13 +14,13 @@ void imshow_250xN(const string& name_, const Mat& patch) {
 
 	cv::Mat bigpatch; cv::Size size = patch.size();
 	float asp = (float)size.height / size.width;
-	float new_size = 550;
-	cv::resize(patch, bigpatch, cv::Size(new_size, new_size * asp));
+	int new_size = 550;
+	cv::resize(patch, bigpatch, cv::Size((int)new_size, int(new_size * asp)));
 	imshow(name_, bigpatch);
 
 }
 
 double getcputime(void) { return (double)clock(); } /// (double)CLOCKS_PER_SEC
-float  mili2sectime(double t) { return ((float)t) / (double)CLOCKS_PER_SEC; }
+float  mili2sectime(double t) { return (float)((t) / (double)CLOCKS_PER_SEC); }
 }
 }
