+++ modules/mcc/src/graph_cluster.hpp
@@ -17,8 +17,8 @@ public:
 	~CB0cluster();
 
 	inline void setVertex(const std::vector<cv::Point> &V) { X = V; }
-	inline void setB0(const std::vector<float> &b0) { B0 = b0;  }
-	inline void setWeight(const std::vector<float> &Weight) { W = Weight; }
+	inline void setB0(const std::vector<double> &b0) { B0 = b0;  }
+	inline void setWeight(const std::vector<double> &Weight) { W = Weight; }
 
 	void group();
 
@@ -28,8 +28,8 @@ private:
 
 	//entrada
 	std::vector<cv::Point> X;
-	std::vector<float> B0;
-	std::vector<float> W;
+	std::vector<double> B0;
+	std::vector<double> W;
 
 	//salida
 	std::vector<int> G;
@@ -41,7 +41,7 @@ private:
 	void find(const std::vector<T> &A, std::vector<int> &indx)
 	{
 		indx.clear();
-		for (size_t i = 0; i < A.size(); i++)
+		for (int i = 0; i < (int)A.size(); i++)
 		if (A[i])indx.push_back(i);
 
 	}
